node: split openflowManager and service health check code
openflowManager code is only marginally health-check related, while the
service code provides healthz responses for local traffic policy
services on the node.

No code changes, just moving stuff around.

(cherry picked from commit c8489e3ff9c321e77f265dc9d484ed2549df4a6b)
(cherry picked from commit 1a2774d35517757f73f7a2d974266f1bd09b203f)
