+++ go-controller/pkg/node/node.go
@@ -815,3 +815,135 @@ func upgradeServiceRoute(bridgeName string) error {
 	}
 	return nil
 }
+
+// checkForStaleOVSInternalPorts checks for OVS internal ports without any ofport assigned,
+// they are stale ports that must be deleted
+func checkForStaleOVSInternalPorts() {
+	// Track how long scrubbing stale interfaces takes
+	start := time.Now()
+	defer func() {
+		klog.V(5).Infof("CheckForStaleOVSInternalPorts took %v", time.Since(start))
+	}()
+
+	stdout, _, err := util.RunOVSVsctl("--data=bare", "--no-headings", "--columns=name", "find",
+		"interface", "ofport=-1")
+	if err != nil {
+		klog.Errorf("Failed to list OVS interfaces with ofport set to -1")
+		return
+	}
+	if len(stdout) == 0 {
+		return
+	}
+	// Batched command length overload shouldn't be a worry here since the number
+	// of interfaces per node should never be very large
+	// TODO: change this to use libovsdb
+	staleInterfaceArgs := []string{}
+	values := strings.Split(stdout, "\n\n")
+	for _, val := range values {
+		klog.Warningf("Found stale interface %s, so queuing it to be deleted", val)
+		if len(staleInterfaceArgs) > 0 {
+			staleInterfaceArgs = append(staleInterfaceArgs, "--")
+		}
+
+		staleInterfaceArgs = append(staleInterfaceArgs, "--if-exists", "--with-iface", "del-port", val)
+	}
+
+	_, stderr, err := util.RunOVSVsctl(staleInterfaceArgs...)
+	if err != nil {
+		klog.Errorf("Failed to delete OVS port/interfaces: stderr: %s (%v)",
+			stderr, err)
+	}
+}
+
+// checkForStaleOVSRepresentorInterfaces checks for stale OVS ports backed by Repreresentor interfaces,
+// derive iface-id from pod name and namespace then remove any interfaces assoicated with a sandbox that are
+// not scheduled to the node.
+func checkForStaleOVSRepresentorInterfaces(nodeName string, wf factory.ObjectCacheInterface) {
+	// Get all ovn-kuberntes Pod interfaces. these are OVS interfaces that have their external_ids:sandbox set.
+	out, stderr, err := util.RunOVSVsctl("--columns=name,external_ids", "--data=bare", "--no-headings",
+		"--format=csv", "find", "Interface", "external_ids:sandbox!=\"\"")
+	if err != nil {
+		klog.Errorf("Failed to list ovn-k8s OVS interfaces:, stderr: %q, error: %v", stderr, err)
+		return
+	}
+
+	// parse this data into local struct
+	type interfaceInfo struct {
+		Name       string
+		Attributes map[string]string
+	}
+
+	lines := strings.Split(out, "\n")
+	interfaceInfos := make([]*interfaceInfo, 0, len(lines))
+	for _, line := range lines {
+		cols := strings.Split(line, ",")
+		// Note: There are exactly 2 column entries as requested in the ovs query
+		// Col 0: interface name
+		// Col 1: space separated key=val pairs of external_ids attributes
+		if len(cols) < 2 {
+			// unlikely to happen
+			continue
+		}
+		ifcInfo := interfaceInfo{Name: strings.TrimSpace(cols[0]), Attributes: make(map[string]string)}
+		for _, attr := range strings.Split(cols[1], " ") {
+			keyVal := strings.SplitN(attr, "=", 2)
+			if len(keyVal) != 2 {
+				// unlikely to happen
+				continue
+			}
+			ifcInfo.Attributes[keyVal[0]] = keyVal[1]
+		}
+		interfaceInfos = append(interfaceInfos, &ifcInfo)
+	}
+
+	if len(interfaceInfos) == 0 {
+		return
+	}
+
+	// list Pods and calculate the expected iface-ids.
+	// Note: we do this after scanning ovs interfaces to avoid deleting ports of pods that where just scheduled
+	// on the node.
+	pods, err := wf.GetPods("")
+	if err != nil {
+		klog.Errorf("Failed to list pods. %v", err)
+		return
+	}
+	expectedIfaceIds := make(map[string]bool)
+	for _, pod := range pods {
+		if pod.Spec.NodeName == nodeName {
+			// Note: wf (WatchFactory) *usually* returns pods assigned to this node, however we dont rely on it
+			// and add this check to filter out pods assigned to other nodes. (e.g when ovnkube master and node
+			// share the same process)
+			expectedIfaceIds[util.GetIfaceId(pod.Namespace, pod.Name)] = true
+		}
+	}
+
+	// Remove any stale representor ports
+	for _, ifaceInfo := range interfaceInfos {
+		// ignore non-vf representor ports
+		if _, ok := ifaceInfo.Attributes["vf-netdev-name"]; !ok {
+			continue
+		}
+		ifaceId, ok := ifaceInfo.Attributes["iface-id"]
+		if !ok {
+			klog.Warningf("iface-id attribute was not found for OVS interface %s. "+
+				"skipping cleanup check for interface", ifaceInfo.Name)
+			continue
+		}
+		if _, ok := expectedIfaceIds[ifaceId]; !ok {
+			klog.Warningf("Found stale OVS Interface, deleting OVS Port with interface %s", ifaceInfo.Name)
+			_, stderr, err := util.RunOVSVsctl("--if-exists", "--with-iface", "del-port", ifaceInfo.Name)
+			if err != nil {
+				klog.Errorf("Failed to delete interface %q . stderr: %q, error: %v",
+					ifaceInfo.Name, stderr, err)
+				continue
+			}
+		}
+	}
+}
+
+// checkForStaleOVSInterfaces periodically checks for stale OVS interfaces
+func checkForStaleOVSInterfaces(nodeName string, wf factory.ObjectCacheInterface) {
+	checkForStaleOVSInternalPorts()
+	checkForStaleOVSRepresentorInterfaces(nodeName, wf)
+}
